rag/requirements.txt
fastapi
uvicorn
pydantic
onnxruntime
transformers
numpy

pinecone

PyMuPDF
python-dotenv
psycopg2
supabase